Move h5pContentFactory to H5p module
