+++ apps/server/src/modules/h5p-editor/controller/api-test/h5p-editor-files.api.spec.ts
@@ -7,7 +7,6 @@ import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
 import { HttpStatus, INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { courseFactory } from '@testing/factory/course.factory';
-import { h5pContentFactory } from '@testing/factory/h5p-content.factory';
 import { lessonFactory } from '@testing/factory/lesson.factory';
 import { UserAndAccountTestFactory } from '@testing/factory/user-and-account.test.factory';
 import { TestApiClient } from '@testing/test-api-client';
@@ -16,6 +15,7 @@ import { H5PContent, H5PContentParentType, H5PContentProperties } from '../../en
 import { H5PEditorTestModule } from '../../h5p-editor-test.module';
 import { H5P_CONTENT_S3_CONNECTION, H5P_LIBRARIES_S3_CONNECTION } from '../../h5p-editor.config';
 import { ContentStorage, LibraryStorage, TemporaryFileStorage } from '../../service';
+import { h5pContentFactory } from '../../testing';
 
 const helpers = {
 	buildMetadata(
