+++ apps/server/src/modules/h5p-editor/controller/api-test/h5p-editor-get-player.api.spec.ts
@@ -6,12 +6,12 @@ import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
 import { HttpStatus, INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { cleanupCollections } from '@testing/cleanup-collections';
-import { h5pContentFactory } from '@testing/factory/h5p-content.factory';
 import { lessonFactory } from '@testing/factory/lesson.factory';
 import { UserAndAccountTestFactory } from '@testing/factory/user-and-account.test.factory';
 import { TestApiClient } from '@testing/test-api-client';
 import { H5PEditorTestModule } from '../../h5p-editor-test.module';
 import { H5P_CONTENT_S3_CONNECTION, H5P_LIBRARIES_S3_CONNECTION } from '../../h5p-editor.config';
+import { h5pContentFactory } from '../../testing';
 
 const buildContent = () => {
 	const contentId = new ObjectId(0).toString();
