+++ apps/server/src/modules/h5p-editor/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './h5p-content.factory';
